Logic implimented

+++ script.js
@@ -1,18 +1,26 @@
 function register() {
-  newUser = {
-    id: username.value,
-    name: fname.value,
-    email: email.value,
-    password: password.value,
-    balance: 0,
-  };
-
-  if (localStorage.getItem(username.value)) {
-    alert("Username Exits");
+  if (username.value == "" || fname.value == "" || password.value == "") {
+    errorText.innerHTML = "Please fill the Form!";
   } else {
-    localStorage.setItem(1, JSON.stringify(newUser));
-    alert("Account Created!");
-    window.location = "./dashboard.html";
+    if (confirmPassword.value != password.value) {
+      errorText.innerHTML = "Password Not Match!";
+    } else {
+      newUser = {
+        id: username.value,
+        name: fname.value,
+        email: email.value,
+        password: password.value,
+        balance: 0,
+      };
+
+      if (localStorage.getItem(username.value)) {
+        alert("Username Exits");
+      } else {
+        localStorage.setItem(1, JSON.stringify(newUser));
+        alert("Account Created!");
+        window.location = "./dashboard.html";
+      }
+    }
   }
 }
 
@@ -23,46 +31,69 @@ function logout() {
 }
 
 function login() {
-  textData = "";
-
-  if (!localStorage.getItem(username.value)) {
-    textData = "Username Didnt Exits!";
+  if (username.value == "" || password.value == "") {
+    errorText.innerHTML = "Please fill the Form!";
   } else {
-    account = JSON.parse(localStorage.getItem(username.value));
-    console.log(account.password, password.value);
+    textData = "";
 
-    if (account.password != password.value) {
-      console.log(account);
-      textData = "Incorrect Password!";
+    if (!localStorage.getItem(username.value)) {
+      textData = "Username Didnt Exits!";
     } else {
-      alert("Login Success");
-      window.location = "./dashboard.html";
-    }
-  }
+      account = JSON.parse(localStorage.getItem(username.value));
+      console.log(account.password, password.value);
 
-  errorText.innerHTML = textData;
+      if (account.password != password.value) {
+        console.log(account);
+        textData = "Incorrect Password!";
+      } else {
+        alert("Login Success");
+        window.location = "./dashboard.html";
+      }
+    }
+
+    errorText.innerHTML = textData;
+  }
 }
 
 function deposit() {
   user = JSON.parse(localStorage.getItem(1));
   if (dPassword.value == user.password) {
-    alert("Amount Added!");
-    user.balance += parseInt(dAmmount.value);
-    amountBalance.innerHTML = `${user.balance}`;
-    localStorage.setItem(1, JSON.stringify(user));
+    if (dAmmount.value == "") {
+      dErrorText.innerHTML = `Amount Incorrect!`;
+    } else {
+      alert("Amount Added!");
+      user.balance += parseInt(dAmmount.value);
+      amountBalance.innerHTML = `${user.balance}`;
+      localStorage.setItem(1, JSON.stringify(user));
+    }
   } else {
-    alert("Password Incorrect!");
+    dErrorText.innerHTML = `Password Incorrect!`;
   }
+
+  setTimeout(() => {
+    dErrorText.innerHTML = "";
+  }, 3000);
 }
 
 function withdraw() {
+  amount = parseInt(wAmmount.value);
   user = JSON.parse(localStorage.getItem(1));
-  if (wPassword.value == user.password) {
-    alert("Amount Withdrawn!");
-    user.balance -= parseInt(wAmmount.value);
-    amountBalance.innerHTML = `${user.balance}`;
-    localStorage.setItem(1, JSON.stringify(user));
+  if (wAmmount.value == "") {
+    wErrorText.innerHTML = `Amount Incorrect!`;
+  } else if (amount > user.balance) {
+    wErrorText.innerHTML = `Not Enough Money!`;
   } else {
-    alert("Password Incorrect!");
+    if (wPassword.value == user.password) {
+      alert("Amount Withdrawn!");
+      user.balance -= amount;
+      amountBalance.innerHTML = `${user.balance}`;
+      localStorage.setItem(1, JSON.stringify(user));
+    } else {
+      wErrorText.innerHTML = `Password Incorrect!`;
+    }
   }
+
+  setTimeout(() => {
+    wErrorText.innerHTML = "";
+  }, 3000);
 }
